+++ packages/compiler/src/declarations.d.ts
@@ -35,15 +35,21 @@ declare module 'find-free-port' {
 }
 
 declare module '@nuxt/friendly-errors-webpack-plugin' {
+  import type { Compiler } from 'webpack';
+
   class FriendlyErrorsWebpackPlugin {
     constructor(opts?: Record<string, unknown>);
+    apply(compiler: Compiler): void;
   }
   export = FriendlyErrorsWebpackPlugin;
 }
 
 declare module 'case-sensitive-paths-webpack-plugin' {
+  import type { Compiler } from 'webpack';
+
   class CaseSensitivePathsPlugin {
     constructor(opts?: Record<string, unknown>);
+    apply(compiler: Compiler): void;
   }
   export = CaseSensitivePathsPlugin;
 }
@@ -54,30 +60,42 @@ declare module 'webpack-node-externals' {
 }
 
 declare module 'webpack-bundle-analyzer' {
+  import type { Compiler } from 'webpack';
+
   class BundleAnalyzerPlugin {
     constructor(opts?: Record<string, unknown>);
+    apply(compiler: Compiler): void;
   }
   export { BundleAnalyzerPlugin };
 }
 
 declare module '@statoscope/webpack-plugin' {
+  import type { Compiler } from 'webpack';
+
   class StatoscopeWebpackPlugin {
     constructor(opts?: Record<string, unknown>);
+    apply(compiler: Compiler): void;
   }
 
   export default StatoscopeWebpackPlugin;
 }
 
 declare module 'webpack/lib/FlagDependencyUsagePlugin.js' {
+  import type { Compiler } from 'webpack';
+
   class FlagDependencyUsagePlugin {
     constructor(explanation?: boolean);
+    apply(compiler: Compiler): void;
   }
   export = FlagDependencyUsagePlugin;
 }
 
 declare module 'webpack/lib/optimize/FlagIncludedChunksPlugin.js' {
+  import type { Compiler } from 'webpack';
+
   class FlagIncludedChunksPlugin {
     constructor();
+    apply(compiler: Compiler): void;
   }
   export = FlagIncludedChunksPlugin;
 }